Client and server changes. Added login form to client etc.
